LayoutTests/imported/w3c/web-platform-tests/css/css-values/tree-counting/sibling-index-keyframe-font-weight-dynamic-expected.txt


FAIL Initially, the sibling-index() is 3 for #target assert_equals: expected "300" but got "400"
FAIL Removing a preceding sibling of #target reduces the sibling-index() assert_equals: expected "200" but got "400"
